+++ money_manager/src/components/Spending.js
@@ -48,7 +48,7 @@ class Expense extends Component {
 
   handleAmount = e => {
     this.setState({
-      amount: parseInt(e.target.value)
+      amount: parseFloat(e.target.value)
     });
   };
